+++ src/prompt.ts
@@ -1,100 +1,79 @@
-import * as fs from "fs";
 import * as vscode from "vscode";
 import { Settings } from "./Settings";
-import { getFilesList, PathDetails } from './utils';
+import { BROWSE_FILE, CommandTypes, getFilesList, PathDetails } from './utils';
 
 export class Prompt {
   static getFilePath = async () => {
     const defaultUri = vscode.workspace.workspaceFolders ? vscode.workspace.workspaceFolders[0].uri : undefined;
     const filePaths = await vscode.window.showOpenDialog({
       defaultUri,
-      filters: { type: ["js"] },
-      title: "Please select a javascript file",
+      title: "Please select a file",
     });
     const filePath = filePaths && filePaths[0].fsPath.toString();
     return filePath;
   };
 
-  static pickAScriptFile = async (output: vscode.OutputChannel): Promise<string | undefined> => {
+  static pickACommand = async (scriptPaths: string[] = [], output: vscode.OutputChannel, shouldBrowseFile: boolean = false) => {
     try {
-      output.appendLine("Picking a File...");
+      output.appendLine("Picking a Command...");
 
-      // get list of all files from the given script paths
-      const scriptFiles = Settings.scriptsPath.reduce((res, scriptPath) => res.concat(
+      const scriptFiles = scriptPaths.reduce((res, scriptPath) => res.concat(
         getFilesList(scriptPath)
           .map(file => ({
             ...file,
-            fileName: file.fileName.replace(/( |_|-){1,}/g, ' ').trim()
+            fileName: file.fileName.replace(/( |_|-){1,}/g, ' ').trim(),
+            filePath: file.filePath.replace(/\\/g, "/")
           }))
-          .filter(file => file.extension === ".js")
-      ), [] as PathDetails[]);
+      ), [] as PathDetails[])
+        .filter((currentFile, index, files) => files.findIndex(file => file.filePath === currentFile.filePath) === index); // filter out repetitive path
 
-      // transforming to pick list format
-      const filesPickList: vscode.QuickPickItem[] = scriptFiles.map(file => ({
+      const filesList = scriptFiles.map(file => ({
         label: file.fileName,
         description: file.filePath,
-      }));
-      const BROWSE_FILE = "Browse File";
-      const quickPickList: vscode.QuickPickItem[] = [
-        ...filesPickList,
-        { label: BROWSE_FILE }
-      ];
-      const selectedOption = await vscode.window.showQuickPick(quickPickList, {
-        placeHolder: "Please select any script file",
-      });
-      // return if no file name is selected
-      if (!selectedOption) {
-        output.appendLine(`Selected File : undefined`);
-        return
-      };
+        value: Settings.runFileCommand,
+        filePath: file.filePath,
+        type: CommandTypes.FILE
+      }))
 
-      // if Browse File is selected then browse for a javascript file
-      const scriptPath = selectedOption.label === BROWSE_FILE ? await Prompt.getFilePath() : selectedOption.description;
-
-      output.appendLine(`Selected File : ${scriptPath}`);
-
-      if (fs.existsSync(scriptPath || '')) return scriptPath;
-
-      return;
-    } catch (error: any) {
-      output.appendLine(error.message);
-      return;
-    }
-  };
-
-  static pickACommand = async (scriptFilePath: string, output: vscode.OutputChannel) => {
-    try {
-      output.appendLine("Picking a Command...");
-      // gracefully require script file
-      const commandEntries = (() => {
+      const { functionalCommands, shellCommands } = scriptFiles.reduce((result, file) => {
         try {
-          delete require.cache[require.resolve(scriptFilePath)];
-          const scriptFile = require(require.resolve(scriptFilePath));
-          return Object.entries(scriptFile);
+          delete require.cache[require.resolve(file.filePath)];
+          const scriptFile = require(require.resolve(file.filePath));
+          const commandEntries = Object.entries(scriptFile);
+
+          commandEntries.forEach(([commandName, code]) => {
+            const command = {
+              label: commandName.replace(/( |_|-){1,}/g, ' ').trim(),
+              value: code,
+              description: file.filePath,
+              filePath: file.filePath,
+              type: typeof code === "string" ? CommandTypes.SHELL : CommandTypes.FUNCTIONAL
+            }
+            if (typeof command.value === 'string') result.shellCommands.push(command);
+            if (typeof command.value === 'function') result.functionalCommands.push(command);
+          })
+
+          return result;
         } catch (err) {
           console.log(err);
-          return []
+          return result
         }
-      })()
-        .filter(([_key, val]) => ["string", "function"].includes(typeof val));
+      }, { functionalCommands: [], shellCommands: [] } as any)
 
-
-      // transforming to pick list format
-      const commandsPickList: vscode.QuickPickItem[] = commandEntries.map(([commandName, code]) => ({
-        label: commandName.replace(/( |_|-){1,}/g, ' ').trim(),
-        value: code,
-        description: typeof code === "string" ? "Shell Command" : "(vscode, args, editorProps, output) => { ... }"
-      }));
-
-      const RUN_FILE = "Run File";
-
-      const quickPickList: vscode.QuickPickItem[] = [
-        { label: RUN_FILE },
-        ...commandsPickList,
+      const quickPickList = [
+        { label: "Run File", kind: vscode.QuickPickItemKind.Separator },
+        ...filesList,
+        { label: "Run Functional Commands", kind: vscode.QuickPickItemKind.Separator },
+        ...functionalCommands,
+        { label: "Run Shell Commands", kind: vscode.QuickPickItemKind.Separator },
+        ...shellCommands,
+        { kind: vscode.QuickPickItemKind.Separator }
       ];
 
+      if (shouldBrowseFile) quickPickList.push({ label: BROWSE_FILE, value: BROWSE_FILE })
+
       const selectedCommand: any = await vscode.window.showQuickPick(quickPickList, {
-        placeHolder: 'Please select command to execute',
+        placeHolder: 'Please select command or File to execute',
       });
 
       // return if no command is selected
@@ -103,11 +82,9 @@ export class Prompt {
         return;
       };
 
-      const command = selectedCommand.label === RUN_FILE ? [RUN_FILE, Settings.runFileFormat] : [selectedCommand.label, selectedCommand.value];
+      output.appendLine(`\nSelected Command: ${JSON.stringify(selectedCommand, null, 2)}\n`);
 
-      output.appendLine(`Selected Command: ${[command[0]]}`);
-
-      return command;
+      return selectedCommand;
     } catch (error: any) {
       output.appendLine(error.message);
       return;
@@ -116,8 +93,7 @@ export class Prompt {
 
   static showPopupMessage = (message: string, action: "info" | "warning" | "error", ...args: any) => {
     if (action === "info") {
-      Settings.showInfoMsg &&
-        vscode.window.showInformationMessage(message, ...args);
+      vscode.window.showInformationMessage(message, ...args);
     } else if (action === "error") {
       vscode.window.showErrorMessage(message, ...args);
     } else if (action === "warning") {